set connection in authenticated_connection
This was originally just returning a string and caused everything to
crap itself. I've changed it so that it works. I also included the
necessary access_secret parameter for Rdio authenticated methods to
be able to work.

DRYed up both auth and unauth connect methods to call a connection
initializer method.